+++ src/modules/webauthldap/mod_webauthldap.c
@@ -83,7 +83,7 @@ int sasl_interact_stub(LDAP *ld,
 static const char *
 cfg_str(cmd_parms * cmd, void *mconf, const char *arg)
 {
-    int e = (int)cmd->info;
+    intptr_t e = (intptr_t) cmd->info;
     char *error_str = NULL;
     /* MWAL_DCONF *dconf = (MWAL_DCONF *) mconf; */
 
@@ -121,7 +121,7 @@ cfg_str(cmd_parms * cmd, void *mconf, const char *arg)
         error_str =
             apr_psprintf(cmd->pool,
                          "Invalid value cmd->info(%d) for directive %s",
-                         e, cmd->directive->directive);
+                         (int) e, cmd->directive->directive);
         break;
 
     }
@@ -134,7 +134,7 @@ cfg_str(cmd_parms * cmd, void *mconf, const char *arg)
 static const char *
 cfg_flag(cmd_parms * cmd, void *mconfig, int flag)
 {
-    int e = (int)cmd->info;
+    intptr_t e = (intptr_t) cmd->info;
     char *error_str = NULL;
     /* MWAL_DCONF *dconf = (MWAL_DCONF *) mconfig; */
 
@@ -156,7 +156,7 @@ cfg_flag(cmd_parms * cmd, void *mconfig, int flag)
         error_str =
             apr_psprintf(cmd->pool,
                          "Invalid value cmd->info(%d) for directive %s",
-                         e, cmd->directive->directive);
+                         (int) e, cmd->directive->directive);
         break;
 
     }
@@ -169,7 +169,7 @@ cfg_flag(cmd_parms * cmd, void *mconfig, int flag)
 static const char *
 cfg_multistr(cmd_parms * cmd, void *mconf, const char *arg)
 {
-    int e = (int)cmd->info;
+    intptr_t e = (intptr_t) cmd->info;
     char *error_str = NULL;
     MWAL_DCONF *dconf = (MWAL_DCONF *) mconf;
     char** attrib;
@@ -188,7 +188,7 @@ cfg_multistr(cmd_parms * cmd, void *mconf, const char *arg)
         error_str =
             apr_psprintf(cmd->pool,
                          "Invalid value cmd->info(%d) for directive %s",
-                         e, cmd->directive->directive);
+                         (int) e, cmd->directive->directive);
         break;
 
     }
@@ -198,7 +198,7 @@ cfg_multistr(cmd_parms * cmd, void *mconf, const char *arg)
 static const char *
 cfg_take12(cmd_parms *cmd, void *mconfig, const char *w1, const char *w2)
 {
-    int e = (int)cmd->info;
+    intptr_t e = (intptr_t) cmd->info;
     char *error_str = NULL;
 
     MWAL_SCONF *sconf = (MWAL_SCONF *)
@@ -213,8 +213,8 @@ cfg_take12(cmd_parms *cmd, void *mconfig, const char *w1, const char *w2)
         default:
             error_str = 
                 apr_psprintf(cmd->pool,
-                             "Invalid value cmd->info(%d) for directive %s", e,
-                             cmd->directive->directive);
+                             "Invalid value cmd->info(%d) for directive %s",
+                             (int) e, cmd->directive->directive);
             break;
     }
     return error_str;
